Add: assets, add: helpers, add: music player page, add: theme, update: main
Add assets that contain an image in this case the album image and the song in .mp3.
Add helpers, in this folder there is the lyrics of the song in a method called getLyrics.
Add music player page, this will be the main page of the application, by the moment there is a scaffold with a Hello word in a center widget.
Add theme, in this folder there is a file that contains a theme based in the dark theme with a custom color and with a custom icon theme.
Update main implementing the music player page in the home of the material app.