Add: audioplayer model, update: pubspec, update: main, update: music player page.
Add audioplayer model to manage two properties with the provider, one boolean playing and the animationController to manage the animation of the disc spinning when the play button is on.
Update pubspec adding animate_do and provider package from the pub.dev.
Update main enclosing the material app into a multiprovider to use the provider with its create depends on the AudioPlayerModel.
Update music player page enclosing the disc image into a SpinPerfect to get the visual effect that the disc is spinning, also in the controller adding the animation controller from the AudioPlayerModel to comunicate between widgets, also in the onPressed of the play pause button, implement the controller stop and repeat with the purpose to when the button is on play the controller calls the repeat and when is on pause the controller calls the stop.