Update: pubspec, update: audio player model, update: music player page.
Update pubspec adding asset audio player package from pub.dev.
Update audio player model fixing the getter of porcentaje.
Update music player page adding a new property called firstTime to know if the song is played for first time, this property is initialized in true, also creating an instance of assetAudioPlayer, also creating a method called open to manage the open of the song the currentPosition and the songDuration.
Also implementing the songDuration in the top of the bar progress, the current second in the bottom of the bar progress and using the porcentaje to change the height of the container to show the bar progress depends on the currentSecond of the song.